Homework #2
Implemented recomendations from the code review
