+++ homework2/src/main/java/ru/otus/spring/domain/TextQuestion.java
@@ -1,13 +1,13 @@
 package ru.otus.spring.domain;
 
+import ru.otus.spring.domain.dto.QuestionStatus;
+import ru.otus.spring.services.AnswerFactory;
+import ru.otus.spring.services.QuestionFactory;
+import ru.otus.spring.services.TextAnswerFactory;
+
 public class TextQuestion extends Question {
 
     private static final String ANSWER_INSTRUCTION = "Enter the answer as text.";
-
-    public static TextQuestion createFromStringRepresentation(int questionId, String text, String correctAnswerStringRepresentation) {
-        TextAnswer correctAnswer = TextAnswer.of(correctAnswerStringRepresentation);
-        return new TextQuestion(questionId, text, correctAnswer);
-    }
 
     public TextQuestion(int questionId, String text, TextAnswer correctAnswer) {
         super(questionId, text, null, correctAnswer);
@@ -20,7 +20,20 @@ public class TextQuestion extends Question {
 
     @Override
     public boolean giveAnswer(String answerStringRepresentation) {
-        TextAnswer answer = TextAnswer.of(answerStringRepresentation);
+        AnswerFactory answerFactory = new TextAnswerFactory();
+        Answer answer = answerFactory.createAnswer(answerStringRepresentation);
         return checkAnswerAndSetQuestionStatus(answer);
+    }
+
+    private boolean checkAnswerAndSetQuestionStatus(Answer answer) {
+        givenAnswer = answer;
+        boolean isAnswerCorrect = correctAnswer.isEqual(answer);
+
+        if(isAnswerCorrect) {
+            questionStatus = QuestionStatus.CORRECT_ANSWER;
+        } else {
+            questionStatus = QuestionStatus.WRONG_ANSWER;
+        }
+        return isAnswerCorrect;
     }
 }
