+++ homework2/src/main/java/ru/otus/spring/services/AnswerFactory.java
@@ -1,0 +1,7 @@
+package ru.otus.spring.services;
+
+import ru.otus.spring.domain.Answer;
+
+public interface AnswerFactory {
+    Answer createAnswer(String answerStringRepresentation);
+}
