+++ homework2/src/main/java/ru/otus/spring/services/QuestionFactory.java
@@ -1,0 +1,11 @@
+package ru.otus.spring.services;
+
+import ru.otus.spring.domain.Question;
+import ru.otus.spring.domain.dto.Option;
+import ru.otus.spring.domain.dto.QuestionDto;
+
+import java.util.List;
+
+public interface QuestionFactory {
+    Question createQuestion(QuestionDto questionDto, List<Option> options) throws Exception;
+}
