+++ homework2/src/main/java/ru/otus/spring/services/TextAnswerFactory.java
@@ -1,0 +1,11 @@
+package ru.otus.spring.services;
+
+import ru.otus.spring.domain.Answer;
+import ru.otus.spring.domain.TextAnswer;
+
+public class TextAnswerFactory implements AnswerFactory {
+    @Override
+    public TextAnswer createAnswer(String answerStringRepresentation) {
+        return new TextAnswer(answerStringRepresentation);
+    }
+}
