+++ homework2/src/main/java/ru/otus/spring/utils/FileSerializationUtil.java
@@ -1,0 +1,7 @@
+package ru.otus.spring.utils;
+
+import java.util.List;
+
+public interface FileSerializationUtil {
+    List<Object> serializeFile(String filename, Class type) throws Exception;
+}
